Incorporate recent commits

+++ pkg/stats/stats.go
@@ -1,5 +1,5 @@
 package stats
 
 type Stats struct {
-	WaitCounts map[string]int `json:"waitCounts"`
+	WaitCounts map[string]int64 `json:"waitCounts"`
 }
